autism-related-behaivours-main/requirements.txt
absl-py==2.4.0
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.4.0
kivy-deps.sdl2==0.8.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mediapipe==0.10.32
mpmath==1.3.0
narwhals==2.18.0
networkx==3.6.1
numpy==2.3.5
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.5
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==70.2.0
six==1.17.0
smmap==5.0.3
sounddevice==0.5.5
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
torch==2.10.0+cpu
torchaudio==2.10.0+cpu
torchvision==0.25.0+cpu
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
